Only_Python/Ticket/requirements.txt
alembic==1.4.3
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
asyncqt==0.8.0
attrs==20.2.0
backcall==0.2.0
bleach==3.2.1
certifi==2020.11.8
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cloudipsp==1.0.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.1
entrypoints==0.3
Flask==1.1.2
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
handler==1.3.0
idna==2.10
ipykernel==5.3.4
ipython==7.18.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.1
Mako==1.1.3
MarkupSafe==1.1.1
mastermind==0.6
mistune==0.8.4
nbclient==0.5.0
nbconvert==6.0.3
nbformat==5.0.7
nest-asyncio==1.4.0
notebook==6.1.4
notebook-toggle-code==0.2.1
packaging==20.4
pandocfilters==1.4.2
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.7
pycparser==2.20
pygame==2.0.0.dev8
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
PySide2==5.15.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.1
pywin32==228
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.2
render==1.0.0
requests==2.25.0
Send2Trash==1.5.0
shiboken2==5.15.1
simple-draw==2.6.8
six==1.15.0
SQLAlchemy==1.3.20
sqlparse==0.3.1
telnetlib3==1.0.3
termcolor==1.1.0
terminado==0.9.0
testpath==0.4.4
tornado==6.0.4
traitlets==5.0.4
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
